+++ mondex/adapter-declaration/src/main/kotlin/com/anaplan/engineering/azuki/mondex/adapter/declaration/action/CreateWorldDeclarableAction.kt
@@ -5,7 +5,7 @@ import com.anaplan.engineering.azuki.mondex.adapter.api.CreateWorldBehaviour
 import com.anaplan.engineering.azuki.mondex.adapter.declaration.MondexDeclarationState
 
 open class CreateWorldDeclarableAction(
-    protected val authPurses: Map<String, Pair<Int, Int>>,
+    protected val authPurses: Map<String, Pair<ULong, ULong>>,
     protected val worldName: String = MondexDeclarationState.DEFAULT_WORLD,
 ) : CreateWorldBehaviour(), DeclarableAction<MondexDeclarationState> {
