+++ mondex/adapter-declaration/src/main/kotlin/com/anaplan/engineering/azuki/mondex/adapter/declaration/declaration/WorldDeclaration.kt
@@ -4,12 +4,12 @@ import com.anaplan.engineering.azuki.declaration.Declaration
 
 data class WorldDeclaration(
     override val name: String,
-    val authPurses: Map<String, Pair<Int, Int>>,
+    val authPurses: Map<String, Pair<ULong, ULong>>,
     val operations: List<WorldOperation>,
     override val standalone: Boolean,
 ) : Declaration
 
 sealed interface WorldOperation {
-    data class Transfer(val from: String, val to: String, val value: Int) : WorldOperation
+    data class Transfer(val from: String, val to: String, val value: ULong) : WorldOperation
     data object Ignore : WorldOperation
 }
