+++ mondex/dsl/src/main/kotlin/com/anaplan/engineering/azuki/mondex/dsl/action/PurseActions.kt
@@ -1,5 +1,5 @@
 package com.anaplan.engineering.azuki.mondex.dsl.action
 
 interface PurseActions {
-    fun createPurse(balance: Int, lost: Int)
+    fun createPurse(balance: ULong, lost: ULong)
 }
